requirements.txt
# FastAPI y servidor ASGI
fastapi
uvicorn[standard]

# Base de datos asíncrona con SQLAlchemy y PostgreSQL
sqlalchemy>=1.4
asyncpg

# Autenticación y seguridad
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]

# Manejo de variables de entorno
python-dotenv

# Validación y serialización
pydantic
pydantic[email]
email-validator

# Pruebas
pytest
pytest-asyncio
httpx

# Tipado opcional
types-python-jose
types-passlib

python-multipart
